+++ src/components/components.ts
@@ -1,4 +1,4 @@
-/* Web components. File generated at Tue, 25 Mar 2025 02:33:45 GMT. */
+/* Web components. File generated at Thu, 27 Mar 2025 04:51:10 GMT. */
 export * from './collapsible-content-panel/collapsible-content-panel.component';
 export * from './fixed-scroll-panel/fixed-scroll-panel.component';
 export * from './command-prompt/command-prompt.component';